+++ apps/editor-dev/pages/base-editor.jsx
@@ -1,49 +1,14 @@
 import {
   BaseEditor,
+  InjectView,
   useCommand,
-  useConfig,
-  useEditor,
   useSetConfig,
 } from "@elf-framework/base-editor";
-import { start, useSubscribe } from "@elf-framework/sapa";
+import { start } from "@elf-framework/sapa";
+import { AppLayout, AppLayoutItem } from "@elf-framework/ui";
 import "@elf-framework/ui/style.css";
 
-function InjectView({ components }) {
-  const editor = useEditor();
-  return (
-    <div>
-      injectView ,{" "}
-      {components.map((c) => {
-        const View = editor.getUI(c);
-
-        return <View />;
-      })}
-    </div>
-  );
-}
-
-function SplitLayoutItem({ direction, content }) {
-  return (
-    <div
-      class="split-layout-item"
-      direction={direction}
-      style={{ backgroundColor: "yellow" }}
-    >
-      {content}
-    </div>
-  );
-}
-
-function SplitLayout({ content }) {
-  return <div class="split-layout">{content}</div>;
-}
-
 start(function () {
-  useSubscribe("config:yellow", (value) => {
-    // console.log(value);
-    console.log(useConfig("yellow"));
-  });
-
   return (
     <div>
       <BaseEditor
@@ -85,27 +50,47 @@ start(function () {
                 );
               },
             });
+
+            editor.registerGroupUI("right-panel", {
+              view: () => {
+                return (
+                  <div>
+                    <button
+                      type="button"
+                      onClick={async () => {
+                        const ret = await useCommand("my-command");
+
+                        console.log("return", ret);
+                      }}
+                    >
+                      Sample
+                    </button>
+                  </div>
+                );
+              },
+            });
           },
         ]}
       >
-        <SplitLayout>
-          <SplitLayoutItem direction="top">
-            <h1>Header</h1>
-          </SplitLayoutItem>
-          <SplitLayoutItem direction="left">
-            <h1>Left</h1>
-            <InjectView key="" components={["view"]} />
-          </SplitLayoutItem>
-          <SplitLayoutItem direction="body">
-            <h1>Body</h1>
-          </SplitLayoutItem>
-          <SplitLayoutItem direction="right">
-            <h1>Right</h1>
-          </SplitLayoutItem>
-          <SplitLayoutItem direction="bottom">
-            <h1>Bottom</h1>
-          </SplitLayoutItem>
-        </SplitLayout>
+        <AppLayout>
+          <AppLayoutItem direction="top" height={40}>
+            <InjectView views={["top-panel"]} />
+          </AppLayoutItem>
+          <AppLayoutItem
+            direction="left"
+            width={280}
+            maxWidth={350}
+            resizable={true}
+          >
+            <InjectView views={["left-panel"]} />
+          </AppLayoutItem>
+          <AppLayoutItem direction="right" width={300}>
+            <InjectView groups={["right-panel"]} />
+          </AppLayoutItem>
+          <AppLayoutItem direction="center">
+            <InjectView views={["center-panel"]} />
+          </AppLayoutItem>
+        </AppLayout>
       </BaseEditor>
     </div>
   );