+++ packages/base-editor/src/index.js
@@ -2,3 +2,4 @@ import "@elf-framework/ui/style.css";
 
 export * from "./components/editor/Editor";
 export * from "./components/editor/BaseEditor";
+export * from "./components/InjectView";